add BoltNetwork.PositionAtFrame method
Allows for rewinding the position of individual entities that have a
HitboxBody component.

A use case for this could be to do a lag compensated distance check
between a client entity and target entity. You can rewind both entities
by the travel time (latency) and have a more accurate distance for when
the client actually fired.